requirements.txt
dotenv==0.9.9
et_xmlfile==2.0.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tzdata==2025.2
